+++ src/js/components/slideshow.js
@@ -2,10 +2,11 @@ import Class from '../mixin/class';
 import Slideshow from '../mixin/slideshow';
 import Animations from './internal/slideshow-animations';
 import SliderReactive from '../mixin/slider-reactive';
-import { $$, boxModelAdjust, css, isVisible } from 'uikit-util';
+import SliderPreload from './internal/slider-preload';
+import { boxModelAdjust, css } from 'uikit-util';
 
 export default {
-    mixins: [Class, Slideshow, SliderReactive],
+    mixins: [Class, Slideshow, SliderReactive, SliderPreload],
 
     props: {
         ratio: String,
@@ -21,27 +22,6 @@ export default {
         attrItem: 'uk-slideshow-item',
         selNav: '.uk-slideshow-nav',
         Animations,
-    },
-
-    connected() {
-        if (window.IntersectionObserver) {
-            this.observer = new IntersectionObserver((entries) => {
-                if (entries.some((entry) => entry.isIntersecting)) {
-                    this.preloadSlides();
-                }
-            });
-            this.observer.observe(this.$el);
-        }
-    },
-
-    disconnected() {
-        this.observer && this.observer.disconnect();
-    },
-
-    methods: {
-        preloadSlides() {
-            [1, -1].forEach((i) => removeLazyLoad(this.slides[this.getIndex(this.index + i)]));
-        },
     },
 
     update: {
@@ -62,10 +42,6 @@ export default {
                 height = Math.min(this.maxHeight, height);
             }
 
-            if (isVisible(this.$el)) {
-                this.preloadSlides();
-            }
-
             return { height: height - boxModelAdjust(this.list, 'height', 'content-box') };
         },
 
@@ -75,8 +51,10 @@ export default {
 
         events: ['resize'],
     },
-};
 
-function removeLazyLoad(el) {
-    el && $$('img[loading="lazy"]', el).forEach((el) => (el.loading = ''));
-}
+    methods: {
+        getAdjacentSlides() {
+            return [1, -1].map((i) => this.slides[this.getIndex(this.index + i)]);
+        },
+    },
+};